+++ apps/npm/fmt-darwin-arm64/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @rsvelte/fmt-darwin-arm64
 
+## 0.7.0
+
 ## 0.6.2
 
 ## 0.6.1